Importing now accepts the CSV format generated by the exporter
Also now importing can create AND update pages. To create pages just
pass a CSV without an id column, or make the id field empty for the
row you want to create. If id is present and not empty then the page
with that id will be updated.

Now it's possible to export pages to CSV, make whatever modifications
are necessary and import the same file to update the pages.

Now the importer will display all successes and failures, but will
still only commit results if there are no errors.